Feature/action updates

+++ tools/gh-pages-publish.ts
@@ -1,32 +1,35 @@
-require('dotenv').config()
-const { cd, exec, echo, touch } = require('shelljs')
-const { readFileSync } = require('fs')
-const url = require('url')
+require('dotenv').config();
 
-let repoUrl
-let pkg = JSON.parse(readFileSync('package.json') as any)
+const { cd, exec, echo, touch } = require('shelljs');
+const { readFileSync } = require('fs');
+const url = require('url');
+
+let repoUrl;
+let pkg = JSON.parse(readFileSync('package.json') as any);
 if (typeof pkg.repository === 'object') {
   if (!pkg.repository.hasOwnProperty('url')) {
-    throw new Error('URL does not exist in repository section')
+    throw new Error('URL does not exist in repository section');
   }
-  repoUrl = pkg.repository.url
+  repoUrl = pkg.repository.url;
 } else {
-  repoUrl = pkg.repository
+  repoUrl = pkg.repository;
 }
 
-let parsedUrl = url.parse(repoUrl)
-let repository = (parsedUrl.host || '') + (parsedUrl.path || '')
-let ghToken = process.env.GH_TOKEN
+let parsedUrl = url.parse(repoUrl);
+let repository = (parsedUrl.host || '') + (parsedUrl.path || '');
+let ghToken = process.env.GH_TOKEN;
 
-echo('Deploying docs!!!')
-cd('docs')
-touch('.nojekyll')
-exec('git init')
-exec('git add .')
-exec('git config user.name "dev-warner"')
-exec('git config user.email jwarner@amplience.com"')
-exec('git commit -m "docs(docs): update gh-pages"')
-exec(
-  `git push --force --quiet "https://${ghToken}@${repository}" master:gh-pages`
-)
-echo('Docs deployed!!')
+if (process.env.GITHUB_REF === 'refs/heads/main') {
+  echo('Deploying docs!!!');
+  cd('docs');
+  touch('.nojekyll');
+  exec('git init');
+  exec('git add .');
+  exec('git config user.name "dev-warner"');
+  exec('git config user.email jwarner@amplience.com"');
+  exec('git commit -m "docs(docs): update gh-pages"');
+  exec(
+    `git push --force --quiet "https://${ghToken}@${repository}" master:gh-pages`
+  );
+  echo('Docs deployed!!');
+}
